AUTOCOMPLETE
Used a list of words as source to check for matches with user input
as user types word in textbox.

keyup event was applied to the input element, and when user types in
input, the words list is looped over to check each word to see if what
user typed is contained in any word in the words list, and append
those words to a li tag and added to the ul element.